+++ packages/uploader/src/react/uploader.tsx
@@ -33,6 +33,7 @@ import localeMap from './i18n';
 import { ComponentInstaller } from '@blocklet/ui-react';
 import mime from 'mime-types';
 import xbytes from 'xbytes';
+import Cookie from 'js-cookie';
 
 // Don't forget the CSS: core and the UI components + plugins you are using.
 import '@uppy/core/dist/style.min.css';
@@ -270,9 +271,17 @@ function initUploader(props: any) {
     // docs: https://github.com/tus/tus-js-client/blob/main/docs/api.md
     withCredentials: true,
     endpoint: uploaderUrl,
+
     async onBeforeRequest(req, file) {
       // @ts-ignore
       const { hashFileName, id, meta } = file;
+
+      // @ts-ignore
+      const mockResponse = currentUppy.getFile(id)?.mockResponse || null;
+      if (req.getMethod() === 'PATCH' && mockResponse) {
+        // mock response to avoid next step upload
+        req.send = () => mockResponse;
+      }
 
       const ext = getExt(file);
 
@@ -291,6 +300,11 @@ function initUploader(props: any) {
           return value;
         })
       );
+      // add csrf token if exist
+      const csrfToken = Cookie.get('x-csrf-token');
+      if (csrfToken) {
+        req.setHeader('x-csrf-token', csrfToken);
+      }
 
       // @ts-ignore get folderId when upload using
       const componentDid = window?.uploaderComponentId || window?.blocklet?.componentId;
@@ -333,27 +347,47 @@ function initUploader(props: any) {
         });
       }
 
+      const file = currentUppy.getFile(result.headers['x-uploader-file-id']);
+
+      // @ts-ignore
+      if (req.getMethod() === 'PATCH' && file.mockResponse) {
+        // mock response do nothing
+        return;
+      }
+
       // only call onUploadFinish if it's a PATCH / POST request
       if (['PATCH', 'POST'].includes(result.method) && [200, 500].includes(result.status)) {
         const isExist = [true, 'true'].includes(result.headers['x-uploader-file-exist']);
         const uploadURL = getUrl(result.url, result.headers['x-uploader-file-name']); // upload URL with file name
-        const file = currentUppy.getFile(result.headers['x-uploader-file-id']);
 
         result.file = file;
         result.uploadURL = uploadURL;
+
+        const responseResult = {
+          uploadURL,
+          ...result,
+        };
+
+        currentUppy.setFileState(file.id, {
+          responseResult,
+        });
 
         // exist but not upload
         if (isExist && file) {
           // if POST method check exist
           if (result.method === 'POST') {
-            // pause first,  that not trigger PATCH request
-            currentUppy.pauseResume(file.id);
+            // set mock response to avoid next step upload
+            currentUppy.setFileState(file.id, {
+              mockResponse: res,
+            });
           }
 
           // only trigger uppy event when exist
-          currentUppy.emit('upload-success', file, {
-            uploadURL,
+          currentUppy.emit('upload-success', currentUppy.getFile(file.id), {
+            ...responseResult,
+            body: result.data,
           });
+          currentUppy.emit('postprocess-complete', currentUppy.getFile(file.id));
 
           // @ts-ignore
           currentUppy.calculateTotalProgress();
@@ -373,15 +407,25 @@ function initUploader(props: any) {
       }
 
       const uploadProgressDone = currentUppy.getState().totalProgress === 100;
+
       const shouldAutoCloseAfterDropUpload = currentUppy.getFiles().every((item: any) => item.source === 'DropTarget');
 
       // close uploader when upload progress done and all files are from DropTarget
-      if (uploadProgressDone && shouldAutoCloseAfterDropUpload) currentUppy.close();
+      if (uploadProgressDone && shouldAutoCloseAfterDropUpload) {
+        currentUppy.close();
+      }
     },
     ...tusProps,
   });
-
   // .use(GoldenRetriever);
+
+  currentUppy.on('upload', ({ fileIDs, id }: { fileIDs: string[]; id: string }) => {
+    fileIDs.forEach((fileId: any) => {
+      currentUppy.setFileState(fileId, {
+        uploadID: id,
+      });
+    });
+  });
 
   // add drop target
   if (dropTargetProps) {
@@ -504,6 +548,7 @@ const Uploader = forwardRef((props: UploaderProps & IframeHTMLAttributes<HTMLIFr
   );
 
   useEffect(() => {
+    setPrefixPath(apiPathProps);
     // @ts-ignore
     state.uppy = initUploader({
       ...props,
@@ -617,10 +662,6 @@ const Uploader = forwardRef((props: UploaderProps & IframeHTMLAttributes<HTMLIFr
       })
   );
 
-  useEffect(() => {
-    setPrefixPath(apiPathProps);
-  }, [apiPathProps]);
-
   const Wrapper = popup ? Backdrop : Fragment;
   const wrapperProps = popup
     ? {
@@ -665,6 +706,7 @@ const Uploader = forwardRef((props: UploaderProps & IframeHTMLAttributes<HTMLIFr
         />
         {/* ignore backdrop trigger */}
         <Box
+          className="uploader-container"
           key="uploader-container"
           id={target}
           onClick={(e: any) => e.stopPropagation()}
@@ -736,28 +778,30 @@ const Uploader = forwardRef((props: UploaderProps & IframeHTMLAttributes<HTMLIFr
               },
             },
           }}>
-          <IconButton
-            aria-label="close"
-            onClick={close}
-            sx={{
-              color: '#fafafa',
-              position: 'absolute',
-              ...(isMobile
-                ? {
-                    bottom: `calc(0px - ${closeIconSize} - 16px)`,
-                    left: `calc(50vw - ${closeIconSize} - 8px)`,
-                  }
-                : {
-                    right: `calc(0px - ${closeIconSize} - 16px)`,
-                    top: -12,
-                  }),
-            }}>
-            <CloseIcon
+          {popup && (
+            <IconButton
+              aria-label="close"
+              onClick={close}
               sx={{
-                fontSize: closeIconSize,
-              }}
-            />
-          </IconButton>
+                color: '#fafafa',
+                position: 'absolute',
+                ...(isMobile
+                  ? {
+                      bottom: `calc(0px - ${closeIconSize} - 16px)`,
+                      left: `calc(50vw - ${closeIconSize} - 8px)`,
+                    }
+                  : {
+                      right: `calc(0px - ${closeIconSize} - 16px)`,
+                      top: -12,
+                    }),
+              }}>
+              <CloseIcon
+                sx={{
+                  fontSize: closeIconSize,
+                }}
+              />
+            </IconButton>
+          )}
           {/* @ts-ignore */}
           {state.uppy && (
             <Dashboard