+++ packages/uploader/CHANGELOG.md
@@ -1,16 +1,44 @@
-## 0.1.29 (九月 04, 2024)
+## 0.1.36 (September 14, 2024)
+
+- chore: add x-csrf-token into request headers
+
+## 0.1.35 (September 13, 2024)
+
+- chore: add peer-deps to deps
+
+## 0.1.34 (September 11, 2024)
+
+- chore: add uploader responseResult in sdk
+
+## 0.1.33 (September 11, 2024)
+
+- chore: polish CHANGELOG date i18n
+
+## 0.1.32 (September 11, 2024)
+
+- refactor: polish checking exist file logic
+
+## 0.1.31 (September 09, 2024)
+
+- chore: remove useless close button when popup is false
+
+## 0.1.30 (September 09, 2024)
+
+- fix: resolve the bug of setPrefixPath not working
+
+## 0.1.29 (September 04, 2024)
 
 - feat: support to prevent zip bomb attack in uploader
 
-## 0.1.28 (八月 29, 2024)
+## 0.1.28 (August 29, 2024)
 
 - chore: bump deps to latest
 
-## 0.1.27 (八月 24, 2024)
+## 0.1.27 (August 24, 2024)
 
 - chore(ci): polish github action to avoid ci stop
 
-## 0.1.26 (八月 24, 2024)
+## 0.1.26 (August 24, 2024)
 
 - chore: polish ci
 
@@ -225,7 +253,7 @@
 - chore: update deps
 - chore: adjust uploader provider to use createPortal
 
-## 0.0.53 (十二月 14, 2023)
+## 0.0.53 (December 14, 2023)
 
 - feat: add remove sdk
 
@@ -257,11 +285,11 @@
 
 - chore: update deps
 
-## 0.0.45 (十一月 20, 2023)
+## 0.0.45 (November 20, 2023)
 
 - chore: set generate image number in dall-e 3
 
-## 0.0.44 (十一月 20, 2023)
+## 0.0.44 (November 20, 2023)
 
 - chore: support select model
 
@@ -273,7 +301,7 @@
 
 - chore: update deps
 
-## 0.0.40 (十一月 15, 2023)
+## 0.0.40 (November 15, 2023)
 
 - chore: support dalle3